pyproject.toml
[project]
name = "json_schemas"
version = "0.1.0"
description = "json-schemas using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "snowflake-connector-python>=3.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
kickoff = "json_schemas.main:kickoff"
run_crew = "json_schemas.main:kickoff"
plot = "json_schemas.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
